Downgrade Compile Version

+++ Projects/Exchange/PlexRouter/contracts/IWETH.sol
@@ -1,9 +1,8 @@
-//SPDX-License-Identifier: MIT
-
+// SPDX-License-Identifier: GPL-3.0
 pragma solidity >=0.5.0;
 
 interface IWETH {
     function deposit() external payable;
-    function transfer(address to, uint value) external returns (bool);
-    function withdraw(uint) external;
+    function transfer(address to, uint256 value) external returns (bool);
+    function withdraw(uint256) external;
 }
